Time: 3 ms (85.23%), Space: 11.2 MB (29.79%) - LeetHub

+++ 33-search-in-rotated-sorted-array/33-search-in-rotated-sorted-array.cpp
@@ -10,7 +10,7 @@ class Solution {
     }
 public:
     int search(vector<int>& nums, int target) {
-        int n = (int)nums.size();
+        int n = nums.size();
         if(n == 1){
             return (target == nums[0] ? 0 : -1);
         }
